refactor: url mapping 수정
@RequestMapping 제거
