+++ BE-MyCarMaster/src/main/java/softeer/be_my_car_master/api/engine/controller/EngineController.java
@@ -8,7 +8,6 @@ import org.springdoc.api.annotations.ParameterObject;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import io.swagger.v3.oas.annotations.Operation;
@@ -18,13 +17,12 @@ import softeer.be_my_car_master.api.engine.dto.request.GetEnginesRequest;
 import softeer.be_my_car_master.api.engine.dto.request.GetUnselectableOptionsByEngineRequest;
 import softeer.be_my_car_master.api.engine.dto.response.GetEnginesResponse;
 import softeer.be_my_car_master.api.engine.dto.response.GetUnselectableOptionsByEngineResponse;
-import softeer.be_my_car_master.api.engine.usecase.GetEnginesUseCase;
 import softeer.be_my_car_master.api.engine.usecase.GetUnselectableOptionsByEngineUseCase;
+import softeer.be_my_car_master.api.engine.usecase.get_engines.GetEnginesUseCase;
 import softeer.be_my_car_master.global.exception.BindingParamException;
 import softeer.be_my_car_master.global.response.Response;
 
 @RestController
-@RequestMapping("/engines")
 @RequiredArgsConstructor
 @Tag(name = "Engine", description = "Engine API Document")
 public class EngineController {
@@ -32,7 +30,7 @@ public class EngineController {
 	private final GetEnginesUseCase getEnginesUseCase;
 	private final GetUnselectableOptionsByEngineUseCase getUnselectableOptionsByEngineUseCase;
 
-	@GetMapping
+	@GetMapping("/engines")
 	@Operation(summary = "트림에 따른 엔진 목록을 반환합니다")
 	public Response<GetEnginesResponse> getEngines(
 		@Valid @ParameterObject GetEnginesRequest getEnginesRequest,
@@ -47,7 +45,7 @@ public class EngineController {
 		return Response.createSuccessResponse(getEnginesResponse);
 	}
 
-	@GetMapping("/{engineId}/unselectable-options")
+	@GetMapping("/engines/{engineId}/unselectable-options")
 	@Operation(summary = "엔진 변경 시도시 기존에 선택된 옵션들 중 변경하려는 엔진에서 선택 불가능한 옵션 목록을 반환합니다.")
 	public Response<GetUnselectableOptionsByEngineResponse> getUnselectableOptionsByEngine(
 		@PathVariable Long engineId,
